requirements.txt
gradio
groq
requests
flask
streamlit
python-dotenv